feat(presenter): add title and description fields to list output
The UI "list" page will need to show the human-friendly title and
description fields.  They are not currently included in the list
output, necessitating /n/ additional API queries to retrieve the
full data for each domain record.

To avoid additional queries, add the title and description fields to
the list output.