router: Replace NewApiServiceValidator with RequestResponseValidator
This is more efficient than having an additional middleware as we have
to setup routing.  This is also something which should be improved
in future.

The response validation is disabled by default, because several
responses currently cause problems (either defects in the API spec,
or bugs in the implementation - probably a mix of both!)