refactor: user update operation
This refactor the user update operation to be aligned with the new API changes,
so only the specific fields can be updated, and no more. Furthermore it adds a
specific repository part for this partial update of the resource; some fixes
were found and updates during this refactor.

Thanks to @cheimes for many of the suggestions during the code review.

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>