build: simplify go tools dependencies
This change add a new way to keep the dependencies on the tools used in the
repository. Instead of create a new rule in the makefile, all the dependencies
for the tools are kept in the `tools/` directory, by using a tools.go file
(which is using the 'tools' build tag to avoid to interfere with the repository
module), and it is added a generic rule which for each entry in the `tools.go`
file, it executes a `go install ...` command, generating the tool binary in our
`bin/` directory of the repository.

The specific versions the tools are attached in the `tools/go.mod` file.

See `tools/README.md` file for additional details.

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>