add token expiration test
The current token expiration test uses a token with exp = 0.  This
is trivially expired, as long as the "now" you are comparing against
is > 0.

Add another test, that generates a new token with a time-to-live of
zero.  Then verify it.  Because we validate with nanosecond
resolution, enough time will have elapsed to ensure that the token
is expired.