HMS-1688 fix: populate scripts
Populate scripts are not adapted to the new API changes and new token. This
update fix the scripts to make them works again for local dev environment and
when reaching out the ephemeral environment, so that we can get some test data
populated quickly. This is specially useful for testing the listDomain view.

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>