HMS-2695 feat: JWK encryption with AES-GCM
Implement helpers to encrypt and decrypt a JWK with AES-GCM AEAD.

`EncryptJWK()` takes a raw AES encryption key plus `jwk.Key` and returns
the encrypted key. The key is encrypted and protected with AES-GCM. The
random nonce is pre-pended to the output stream.

`DecryptJWK()` is the reverse operation. It takes a raw AES key and
cipher text with nonce pre-pended. On success, it returns a `jwk.Key`.
Invalid key and tampering are detected by AEAD.

Also pulls the latest `lestrrat-go/jwx@v2` go module.

Signed-off-by: Christian Heimes <cheimes@redhat.com>