HMS-2694 feat: Host conf token MVP
Implement minimal host conf token and signing JWK.

A private ECDSA key is genenerated from a hard-coded seed and the JWK is
stored in memory. In the future, the ECDSA key will be generated from a
random source and the JWK stored in the database.

The `GET /signing_keys` endpoint returns the public JWK.

The `POST /host-conf` endpoint now returns a signed token.

Signed-off-by: Christian Heimes <cheimes@redhat.com>