HMS-2714 fix description and title validation
The validation regexes for the description and title fields are too
restrictive.  Reasonable values are rejected.

Instead, use callback validators to check that the string values
(which are byte sequences) are valid UTF-8, and they do not contain
control characters (except the description field allows Carriage
Return, Line Feed, and Horizontal Tab).