HMS-2837 fix: middleware validation error response
We should not return a plaintext error response as the error won't be displayed
in this case. By returning a new HTTP error the default error handler of the
echo server will take care of it creating the error response in the correct
format. As this is a REST API, the error will be returned in json format.

Pair-Programmed-With: Alejandro Visiedo <avisiedo@redhat.com>