refactor: clarify update authorization check
Rename the function `isSubscriptionManagerIDAuthorizedToUpdate` to
`ensureSubscriptionManagerIDAuthorizedToUpdate`.  The old name makes
it seems like it would return a `bool`, when in fact it return `nil`
on success or an error if there is a problem.

Also convert it from an object method to a bare function, because
the `application` was unused.