compose: use correct account for healthcheck
The compose currently attempts to connect as the container user
"root".  This role does not exist in the postgresql server.  The
liveness test does succeed, but it fills the "log" (stdout) with
annoying messages:

    ... FATAL:  role "root" does not exist

`pg_isready` recognises the standard *libpq* environment variables:
https://www.postgresql.org/docs/16/libpq-envars.html.  Set the
`PGDATABASE`, `PGUSER` and `PGPASSWORD` environment variables in the
database container environment to ensure that the connect test uses
a valid account.  As a result, the log output no longer contains the
the above message every few seconds.

Also add a comment to explain `deployments/backend-compose.yaml` is
used by ipa-hcc CI (it is not referenced elsewhere in this repo,
which caused some confusion).