HMS-1832 refactor: enhance UpdateAgent operation
Initial refactor to enhance UpdateAgent by using GORM operations to avoid the
delete all and create objects again.

Delegate part of the process on GORM and add additional process to remove
reminder certs, servers and locations records.

When the new has more records, it is not a situation because INSERT operation
happens into the database; when the same number of related items, everything
will be UPDATE statements; when less related records than the original, the
remainder records are collected and a DELETE statement operation is launched
only for the necessary records.

Additionally clean-up some old code for the old PartialUpdate method.

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>