HMS-3158 build: split tool dependencies and exlude
This change split the tool dependencies and exclude a dead repository; it is
returned because we got some situation analyzing the dependencies, and I thought
it would be easier when we are not mixing the service dependencies with the
tools.

@cryptomilk detected a dead repository and analyzing the dependencies I
suggested to do this change, to make life easier when facing similar situations
(at least we directly discard what is evoked by the service and the tools).

The dead repository is:

github.com/mitchellh/osext v0.0.0-20151018003038-5e2d6d41470f

At the beginning was removed the dependency on 'goda', because some trial and
error evoked that tool as one of the failing on the dependency; even with this,
it was failing 'go mod tidy' and 'go mod download'.

Additionally it has been added the dead repository as 'exclude', which avoid the
situation when running 'go mod tidy' or 'go mod download'.

Now everything is working without remove any tool, which make me think that some
updates happened on intermediate repositories that fixed the situation.

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>