refactor(HMS-3527): rbac mock as standalone service
Decouple the rbac mock from the main service so a
standalone binary is generated only for the mock,
and remove ENV_NAME from different locations. Finally
it remove not required statements for the mock-rbac
container image.

When using local config.yaml, the app.name should be
set to "idmsvc" value; when it is deployed in a cluster,
the template set the APP_NAME value and inject as environment
variable wherever it is necessary.

This value is used for providing the application parameter
when requesting the ACL to the RBAC service.

Use the appName to know when to return the predefined list
of permissions; for any other requests, it returns the wildcard
combination '*:*:*'.

Add the launcher to debug the rbac mock from vscode.

Initialize logs and add log middleware for the
rbac mock, so we have traces about the behavior.

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>
Co-authored-by: Christian Heimes <cheimes@redhat.com>