test(HMS-3968): update test scripts
This change add support to the manual scripts that allow
to interact with the API locally and ephemeral.

XRHID_USE= indicate which type to use; currently the values
could be "user", "system", "service-account". By default,
depending on the operation will be "user" or "system".

To launch them:

XRHID_USE=service-account ./test/scripts/local-domains-token.sh

https://issues.redhat.com/browse/HMS-3803

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>