doc(HMS-3968): update dev documentation
Update documentation to provide how to use RBAC profiles
and exchange between XRHID profiles when launching manual
requests.

Update the integration test example to be more aligned with
a best way to leverage the wrapper which reduce boilerplate
and is more clear.

Minor change to indicate the Fedora version to install vscode.

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>