refactor(HMS-3968): extend As(RBAC...) semantic
This change extend the current `As` semantic by considering
the RBAC profile to be used. This is, depending on the type
we are passing, and for this specific semantic when an RBAC
profile is specified, it tells to the mock rbac which permissions
are returned. The arguments that we can use for this extended
semantic are:

- RBACSuperAdmin: tell to the rbac mock to set super admin
  permissions.
- RBACAdmin: tell to the rbac mock to set the admin permissions.
- RBACReadOnly: tell to the rbac mock to set the read only
  permissions.
- RBACNoPerms: tell to the rbac mock to return empty list
  of permissions.

If several profiles are indicated, the order matters, and
they are applied from the left to the right.

https://issues.redhat.com/browse/HMS-4280

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>