fix: make vet violation because no deferred call
The call to `time.Since(start).Seconds()` was not
being really deferred. To avoid the situation we
wrap the invocation in a deferred wrapped function
so all the statement is executed at the expected
moment.

This means the `Observe` method now is going to be
valled with the right value of the duration processing
the call.

Indeed this does not differ too much, becouse it was
happening after invoking the `next(c)` invokation but
it is a more correct code.

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>