chore: grafana setup
adds make file similar to Prometheus make commands for starting Grafana
locally in a container.

Both Prometheus and Grafana containers are changed to use bridge network
called `idmsvc_backend`. The reason is for compatibility with MacOS where
Podman host network doesn't work.

To reach idmsvc_backend that is running locally on host, the Prometheus
needs to use `host.containers.internal` which is a special Podman DNS
name for host machine.