fix(HMS-1806): closing of http response body in Pendo client
In one call the body was not closed and in another it was not closed
when HTTP status was other than OK.

Now, in both usages, it is "deferred" closed right after checking that
response did not error-out (and thus `.Body` should be present).

Note: `// nolint:errcheck` is used as linter complains but the code
used is the one mentioned in official docs of `net/http` package.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>