test: MockServer config for pendo mock
With this change, `make run` will start the 'service' with Pendo
client configured against the MockServer. Thus exercising the
pendo_client code.

The "inventory" expectations are removed as they did not work as they
are in wrong format(yaml) to be used in initializers. I.e. it's a cleanup.

This change is also introducing a more general directory structure where
it is expected to put a file named "$PREFIX-expectations.json" into
"test/mockserver" directory. The MockServer will automatically load it.

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>