fix: avoid panic due to unset XRHID on invalid path
If the HTML resource path is unrecognised, the EnforceIdentity
middlewares all get skipped.  This causes the XRHID field in the
context to be uninitialised.  As a consequence, the RBAC middleware
panics because it expects the field to be set.

To avoid this issue, we create a new middleware which always sets
the XRHID into the request context.  This middleware runs after
the FakeIdentity middleware (if used), and before the
EnforceIdentity and RBAC middlewares.

The EnforceIdentity middleware is simplified as a result; it no
longer needs to parse the XRHID and can read it from the context.