fix(HMS-4789): middleware called when not found
This change fix the situation when a not found path
evoke the chain of middlewares to be called in a group.

It adds a new middleware that check the condition on
the global chain of middlewares, and do an early return,
aoviding the chain of middlewares are called.

As it is called as the last middleware in the global
chain, we can log a message indicating the route that
was not found. The metrics will register the group
route with the suffix "/*".

https://issues.redhat.com/browse/HMS-4789

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>