fix(HMS-4848): db-tool migrate down 1
db-tool had an issue that the `steps` arg was not used in practice and thus
variant with `0` was always run.

There is an issue present in run  `db-tool migrate down 0` which ends
with `panic: database locked`. But in practice the issue is:

```
migration failed: cannot drop table domains because other objects depend
on it, constraint fk_ipas_id__domains_id on table ipas depends on table domains
in line 0
```

I.e. one of the older `down` updates is not correct.

This patch is fixing handling of the `steps` arg to be used and
it adds log statement to not hide error like the above.

In addition it adds new make file target to test it more easily and modifies
vscode launch to use down by one step instead of all (the same in makefile target).

Signed-off-by: Petr Vobornik <pvoborni@redhat.com>