fix(HMS-5066): make config.example.yaml secure by default
Set `app.accept_x_rh_fake_identity: false` in the example config
file (which is copied as-is to the container image).  In this way,
the default configuration will be safe.

Add a comment that this file should have safe/appropriate values for
production use.  Overrides, when necessary, should be effected via
make variables, environiment variables, or modifying the config.

Fixes: https://issues.redhat.com/browse/HMS-5066