doc(HMS-5031): update developer documentation
Update several developer documentation at the
repository; README.md to be aligned with the last
changes and versions, and many files at `docs/dev`
directory to be aligned with last state of the
implementation.

The `docs/dev/00-frameworks-and-libraries.md` file
has been updated, removing double entry for RBAC,
and clean-up.

The `docs/dev/TESTING.md` file has been updated to
consider the last refactor for the unit tests on
the repository layer.

The section for kafka is left as the code for the
infrastructure still is part of the repository.

https://issues.redhat.com/browse/HMS-5031

Signed-off-by: Alejandro Visiedo <avisiedo@redhat.com>